% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_data.R
\name{all_cause_rates}
\alias{all_cause_rates}
\title{All Cause Mortality Rates}
\format{
\code{all_cause_rates} is a data.frame containing
  the following
  \itemize{
    \item{Age:} An integer.
    \item{both:} A numeric. All-cause mortality rate for combined male and female.
    \item{male:} A numeric. Male only all-cause mortality rate.
    \item{female:} A numeric. Female only all-cause mortality rate.
  }
}
\usage{
data(all_cause_rates)
}
\description{
All-cause mortality rates for both male and female, male only, and female
  only for the 1974 birth cohort. Data were taken from the CDC Life Tables.
  Vital Statistics of the United States, 1974 Life Tables, Vol. II, Section 5. 1976.
}
\keyword{datasets}
